pyproject.toml
[project]
name = "myfxbookcopying"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cachetools>=5.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.0",
    "fear-and-greed>=0.4",
    "ipython>=9.4.0",
    "jinja2>=3.1.6",
    "pandas>=2.3.1",
    "plotly>=5.24.0",
    "psycopg2-binary>=2.9.10",
    "pyportfolioopt>=1.5.6",
    "quantstats>=0.0.62",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.42",
    "ta>=0.11.0",
    "telethon>=1.36.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.32.0",
    "yfinance>=0.2.65",
    "boto3>=1.35.0",
    "huggingface_hub>=0.20.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]
[project.optional-dependencies]
dev = [
    "catboost>=1.2.8",
    "imbalanced-learn>=0.12.4",
    "ipykernel>=6.30.0",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "scikit-learn>=1.7.1",
    "tqdm>=4.67.1",
    "xgboost>=3.0.3",
    "yfinance>=0.2.65",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[tool.uv.sources]
imbalanced-learn = { git = "https://github.com/scikit-learn-contrib/imbalanced-learn", rev = "master" }

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
]
